swapped to hashrouter, reworked search page, fixed activity name line break
