+++ public/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8"/>
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1"/>
+    <meta name="viewport" content="initial-scale=1, maximum-scale=1"/>
     <meta name="theme-color" content="#000000"/>
     <meta
             name="description"
