+++ src/assets/css/SearchItem.css
@@ -19,7 +19,6 @@
 .search-item-title {
   min-width: 0;
   display: flex;
-  align-items: center;
   gap: 5px;
 }
 .search-item-image {
@@ -31,10 +30,10 @@
   object-fit: cover;
 }
 .search-item-name {
+  min-width: 0;
   font-family: var(--h2-typeface);
   font-size: var(--h2-size);
   overflow: hidden;
-  white-space: nowrap;
   text-overflow: ellipsis;
 }
 .search-item-sport-icon {
