+++ src/index.js
@@ -1,9 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
+import { BrowserRouter, Routes, Route, Navigate, HashRouter } from "react-router-dom";
 import Layout from "./layouts/Layout";
 import LoginLayout from "./layouts/LoginLayout";
-import SearchLayout from "./layouts/SearchLayout";
 import "./assets/css/index.css";
 import { Login } from "./pages/Login";
 import { Activity } from "./pages/Activity";
@@ -18,12 +17,13 @@ import { YourActivities } from "./pages/YourActivities";
 import { Sports } from "./pages/Sports";
 import { Profiles } from "./pages/Profiles";
 import { EditActivity } from "./pages/EditActivity";
+import { Search } from "./pages/Search";
 
 export const backendUrl = "https://api-activee.onrender.com";
 
 export default function App() {
   return (
-    <BrowserRouter>
+    <HashRouter>
       <Routes>
         <Route path="/" element={<LoginLayout />}>
           <Route exact path="login" element={<Login />} />
@@ -41,10 +41,10 @@ export default function App() {
           <Route exact path="profile/edit" element={<EditProfile />} />
           <Route exact path="settings" element={<Settings />} />
           <Route exact path="settings/profiles" element={<Profiles />} />
+          <Route exact path="search" element={<Search />} />
         </Route>
-        <Route exact path="search" element={<SearchLayout />} />
       </Routes>
-    </BrowserRouter>
+    </HashRouter>
   );
 }
